pyproject.toml
[project]
name = "tiktok-multiagent"
version = "0.1.0"
description = "GRWM AI Video Automation Pipeline — Multi-Agent System"
requires-python = ">=3.11"
dependencies = [
    "firebase-admin>=6.0.0",
    "google-generativeai>=0.8.0",
    "google-genai>=0.5.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "moviepy>=2.0.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
vertex = [
    "google-cloud-aiplatform>=1.60.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]

[tool.uv]
package = false
